Reorganized tests for clarity
